requirements.txt
gradio
pdfplumber
sentence-transformers
scikit-learn
requests
python-dotenv
google-generativeai